Add XNNPACK clamp.
Optimal usage of clamps (ReLU, ReLU6, Hardtanh, or any other operator
that can be represented as a clamp) would be to fuse it with the
previous operator if possible, as committing the results of the previous
operator to memory only to read it back for an operation as simple as a
clamp is bound to be inefficient both in terms of performance and power
consumption.  If we have done our job properly, JIT must have
intercepted and replaced all usages of clamp with the corresponding
used version on mobile, hence enabling the said optimal usage.  Having
said that, it may not always be possible to fuse clamps with the
previous operator, either because (1) such usage is still not supported
in PyTorch, because (2) efficient kernels of said operators that
support fused-mode do not exist, or because (3) we failed to properly
intercept and fuse the operation at JIT time.  Regardless, it can very
well be the case that execution ends up in at::clamp which currently
does not support a vectorized ARM backend, significantly hurting
performance.  This PR intercepts and re-routes clamp operations to
XNNPACK on mobile.

ghstack-source-id: 0842d5d50a69f8ec3ae7933433d4a4e6d11a877f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35896