[TensorExpr] Add dtype to class Buf.
Currently Buf represents underlying storage but it didn't have dtype.
That resulted in specifying dtypes in different places and there was no
mechanism to enforce its consistency: e.g. one could've created a kFloat
expression and use a kInt buffer to store its result. Now we're
centralizing where the logic regarding the storage is located and we can
start enforcing semantics rules.

Follow-ups: we can merge Buffer and BufHandle classes as the former is
now a mere wrapper over the latter.

ghstack-source-id: 99b29089c7cf9677ae1e0f36869815e6e122f099
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36611